+++ src/JuliaBUGS.jl
@@ -51,50 +51,6 @@ include("source_gen.jl")
 
 include("BUGSExamples/BUGSExamples.jl")
 
-# BUGS scalar functions that are documented and users expect to work
-const BUGS_SCALAR_FUNCTIONS = [
-    :abs,
-    :arccos,
-    :arccosh,
-    :arcsin,
-    :arcsinh,
-    :arctan,
-    :arctanh,
-    :cloglog,
-    :cos,
-    :cosh,
-    :cumulative,
-    :cut,
-    :density,
-    :deviance,
-    :equals,
-    :exp,
-    :gammap,
-    :ilogit,
-    :icloglog,
-    :integral,
-    :log,
-    :logfact,
-    :loggam,
-    :logit,
-    :max,
-    :min,
-    :phi,
-    :pow,
-    :probit,
-    :round,
-    :sin,
-    :sinh,
-    :solution,
-    :sqrt,
-    :step,
-    :tan,
-    :tanh,
-    :trunc,
-    :sum,
-    :mean,
-]
-
 function check_input(input::NamedTuple)
     valid_pairs = Pair{Symbol,Any}[]
     for (k, v) in pairs(input)
@@ -190,8 +146,8 @@ function finish_checking_repeated_assignments(
     end
 end
 
-function create_graph(model_def, eval_env)
-    pass = AddVertices(model_def, eval_env)
+function create_graph(model_def, eval_env, eval_module=Main)
+    pass = AddVertices(model_def, eval_env, eval_module)
     analyze_block(pass, model_def)
     pass = AddEdges(pass.env, pass.g, pass.vertex_id_tracker)
     analyze_block(pass, model_def)
@@ -211,6 +167,78 @@ function semantic_analysis(model_def, data)
 end
 
 """
+Manages the allowlist of functions that can be used in @bugs macro expressions.
+Only functions in this allowlist or registered via @bugs_primitive are permitted.
+"""
+const BUGS_ALLOWED_FUNCTIONS = Set{Symbol}()
+
+"""
+    is_function_allowed(func_name::Symbol)
+
+Check if a function is allowed to be used in @bugs expressions.
+"""
+is_function_allowed(func_name::Symbol) = func_name in BUGS_ALLOWED_FUNCTIONS
+
+"""
+    validate_bugs_expression(expr, line_num)
+
+Validate that all function calls in the expression are allowed.
+Throws an error if an unregistered function is found.
+"""
+function validate_bugs_expression(expr, line_num)
+    if expr isa Symbol || expr isa Number
+        return nothing  # Base cases are fine
+    elseif Meta.isexpr(expr, :call)
+        func_name = expr.args[1]
+
+        # Check for qualified function names (e.g., Base.exp, Distributions.Normal)
+        if Meta.isexpr(func_name, :.)
+            qualified_expr = func_name
+            unqualified_name =
+                if Meta.isexpr(qualified_expr, :.) && length(qualified_expr.args) >= 2
+                    # For expressions like Base.exp, extract :exp
+                    qualified_expr.args[2].value
+                else
+                    qualified_expr
+                end
+            error(
+                "Qualified function names are not supported in @bugs. Found $(qualified_expr) at $line_num. " *
+                "To use custom functions, declare them with @bugs_primitive macro. " *
+                "Otherwise, use the unqualified function name `$(unqualified_name)` instead.",
+            )
+        elseif func_name isa Symbol && !is_function_allowed(func_name)
+            error(
+                "Function '$func_name' is not allowed in @bugs at $line_num. " *
+                "To use custom functions, declare them with @bugs_primitive macro.",
+            )
+        end
+        # Recursively validate arguments
+        for arg in expr.args[2:end]
+            validate_bugs_expression(arg, line_num)
+        end
+    elseif Meta.isexpr(expr, :ref)
+        # Validate array indexing expressions
+        for arg in expr.args
+            validate_bugs_expression(arg, line_num)
+        end
+    elseif Meta.isexpr(expr, :block)
+        # Validate block expressions
+        for arg in expr.args
+            if !(arg isa LineNumberNode)
+                validate_bugs_expression(arg, line_num)
+            end
+        end
+    elseif Meta.isexpr(expr, :for)
+        # Validate for loop expressions
+        validate_bugs_expression(expr.args[2], line_num)  # loop body
+    elseif Meta.isexpr(expr, :(=))
+        # For assignments, validate both LHS (in case of array indexing) and RHS
+        validate_bugs_expression(expr.args[1], line_num)
+        validate_bugs_expression(expr.args[2], line_num)
+    end
+end
+
+"""
     compile(model_def, data[, initial_params]; skip_validation=false)
 
 Compile the model with model definition and data. Optionally, initializations can be provided. 
@@ -224,17 +252,18 @@ function compile(
     data::NamedTuple,
     initial_params::NamedTuple=NamedTuple();
     skip_validation::Bool=false,
+    eval_module::Module=@__MODULE__,
 )
     # Validate functions by default (for @bugs macro usage)
     # Skip validation only for @model macro
     if !skip_validation
-        parser.validate_bugs_expression(model_def, LineNumberNode(0))
+        validate_bugs_expression(model_def, LineNumberNode(0))
     end
 
     data = check_input(data)
     eval_env = semantic_analysis(model_def, data)
     model_def = concretize_colon_indexing(model_def, eval_env)
-    g = create_graph(model_def, eval_env)
+    g = create_graph(model_def, eval_env, eval_module)
     nonmissing_eval_env = NamedTuple{keys(eval_env)}(
         map(
             v -> begin
@@ -266,62 +295,27 @@ end
 # end
 
 """
-    @bugs_primitive(expr)
+    register_bugs_function(func_name::Symbol)
 
-Currently, only function defined in the `BUGSPrimitives` module can be used in the model definition. 
-This macro allows the user to register a user-defined function or distribution to be used in the model definition.
-
-Example:
-```julia
-julia> @bugs_primitive function f(x) # function
-    return x + 1
-end
-
-julia> JuliaBUGS.f(1)
-2
-
-julia> @bugs_primitive d(x) = Normal(0, x^2) # distribution
-
-julia> JuliaBUGS.d(1)
-Distributions.Normal{Float64}(μ=0.0, σ=1.0)
-```
+Register a function to be allowed in @bugs expressions.
+Used by @bugs_primitive macro.
 """
-macro bugs_primitive(expr)
-    def = MacroTools.splitdef(expr)
-    func_name = def[:name]
-    func_expr = MacroTools.combinedef(def)
-    return quote
-        @eval JuliaBUGS begin
-            # export $func_name
-            $func_expr
-        end
-    end
+function register_bugs_function(func_name::Symbol)
+    push!(BUGS_ALLOWED_FUNCTIONS, func_name)
 end
 
 """
-    @bugs_primitive(func)
-
-`@bugs_primitive` can also be used to register function without definition.
-
-Example
-```julia
-julia> f(x) = x + 1
-
-julia> @bugs_primitive(f)
-
-julia> JuliaBUGS.f(1)
-2
-```
+Helper function to generate the expression for registering a single bugs primitive
 """
-macro bugs_primitive(func::Symbol)
+function _bugs_primitive_expr(func::Symbol, esc_func)
     return quote
-        local f = $(esc(func))
+        local f = $esc_func
         # Check if it's callable by checking if it has methods
         if length(methods(f)) == 0
             error("@bugs_primitive: $($(QuoteNode(func))) is not callable")
         end
         # Add to the allowed functions set
-        JuliaBUGS.parser.register_bugs_function($(QuoteNode(func)))
+        JuliaBUGS.register_bugs_function($(QuoteNode(func)))
         # Also add to JuliaBUGS module for direct access (if not already defined)
         if !isdefined(JuliaBUGS, $(QuoteNode(func)))
             Core.eval(JuliaBUGS, Expr(:const, Expr(:(=), $(QuoteNode(func)), f)))
@@ -329,27 +323,18 @@ macro bugs_primitive(func::Symbol)
         nothing
     end
 end
+
+"""
+    @bugs_primitive(func)
+
+`@bugs_primitive` can also be used to register function without definition.
+"""
+macro bugs_primitive(func::Symbol)
+    return _bugs_primitive_expr(func, esc(func))
+end
+
 macro bugs_primitive(funcs::Vararg{Symbol})
-    exprs = []
-    for func in funcs
-        push!(
-            exprs,
-            quote
-                local f = $(esc(func))
-                # Check if it's callable by checking if it has methods
-                if length(methods(f)) == 0
-                    error("@bugs_primitive: $($(QuoteNode(func))) is not callable")
-                end
-                # Add to the allowed functions set
-                JuliaBUGS.parser.register_bugs_function($(QuoteNode(func)))
-                # Also add to JuliaBUGS module for direct access (if not already defined)
-                if !isdefined(JuliaBUGS, $(QuoteNode(func)))
-                    Core.eval(JuliaBUGS, Expr(:const, Expr(:(=), $(QuoteNode(func)), f)))
-                end
-                nothing
-            end,
-        )
-    end
+    exprs = [_bugs_primitive_expr(func, esc(func)) for func in funcs]
     return Expr(:block, exprs...)
 end
 
@@ -371,24 +356,62 @@ include("serialization.jl")
 include("experimental/ProbabilisticGraphicalModels/ProbabilisticGraphicalModels.jl")
 
 function __init__()
-    # Initialize the allowed functions set with BUGSPrimitives exports
-    empty!(parser.BUGS_ALLOWED_FUNCTIONS)
+    empty!(BUGS_ALLOWED_FUNCTIONS)
 
-    # Add all exported functions from BUGSPrimitives
     for name in names(BUGSPrimitives; all=false)
         if isdefined(BUGSPrimitives, name)
-            push!(parser.BUGS_ALLOWED_FUNCTIONS, name)
+            push!(BUGS_ALLOWED_FUNCTIONS, name)
         end
     end
 
-    # Add BUGS scalar functions
-    for func in BUGS_SCALAR_FUNCTIONS
-        push!(parser.BUGS_ALLOWED_FUNCTIONS, func)
+    for func in [
+        :abs,
+        :arccos,
+        :arccosh,
+        :arcsin,
+        :arcsinh,
+        :arctan,
+        :arctanh,
+        :cloglog,
+        :cos,
+        :cosh,
+        :cumulative,
+        :cut,
+        :density,
+        :deviance,
+        :equals,
+        :exp,
+        :gammap,
+        :ilogit,
+        :icloglog,
+        :integral,
+        :log,
+        :logfact,
+        :loggam,
+        :logit,
+        :max,
+        :min,
+        :phi,
+        :pow,
+        :probit,
+        :round,
+        :sin,
+        :sinh,
+        :solution,
+        :sqrt,
+        :step,
+        :tan,
+        :tanh,
+        :trunc,
+        :sum,
+        :mean,
+    ]
+        push!(BUGS_ALLOWED_FUNCTIONS, func)
     end
 
     # Add basic operators
     for op in [:+, :-, :*, :/, :^, :~, :>, :<, :>=, :<=, :(==), :!, :(:)]
-        push!(parser.BUGS_ALLOWED_FUNCTIONS, op)
+        push!(BUGS_ALLOWED_FUNCTIONS, op)
     end
 end
